mlx5dr: Cmd, Always set reparse bit on RTC creation
Currently we don't have dynamic support over the reparse bit,
setting the bit depends on the actions (e.g.g encap).
The idea is to have 2 RTC over the same STE range one with
reparse on and one with reparse off. But for now we will
pay the performance pain and always set.

Signed-off-by: Alex Vesker <valex@nvidia.com>