common/mlx5: fix RQ size configuration in QP create
The number of WQEBBs was provided to QP create, and QP size was calculated
by multiplying the number of WQEBBs by 64, which is the send WQE size.
When creating RQ in the QP (i.e., vdpa driver), the queue size was bigger
because the receive WQE size is 16.
Provide queue size to QP create instead of the number of WQEBBs.

Fixes: 0c5c4c76151a ("common/mlx5: share DevX QP operations")

Signed-off-by: Raja Zidane <rzidane@nvidia.com>