mlx5dr: Matcher, Support rule based and tables based insertion
Rule based:
The user provides the number of rules that will be used in the
matcher, internally the SW will allocate the amount or resources
sufficient for the given amount of rules.
Internally in rule based mode we will concatenate another smaller
collision matcher and will retry the rule insertion in case of
failure.

Table based:
User provides Row * Col of the hash table, failures can happen
due to resource depletion.

Signed-off-by: Alex Vesker <valex@nvidia.com>