pkg: only build needed drivers
Because RPMs are only for internal usage, only some PMDs are needed.

Add one internal shell function in %prep stage which will find all
drivers which are not to be enabled and the results are feed to meson
--disable_drivers parameter.

This internal shell function is POSIX SH compatible in order to have
most portibility across different build environment.

Signed-off-by: Xiaoyu Min <jackmin@nvidia.com>