crypto/mlx5: fix max number of QPs capability
An indirect mkey is created for each descriptor in a QP, number of
descriptors per QP is configured by the user on QP setup callback.
In mlx cryptodev autotest, the max number of QPs (provided by the driver)
is created, and due to mkey resource limits, QPs creation fail which leads
to the test failing.
Since there is no capability of max number of descriptors provided to
the user, we can't give an exact number of max QPs available.
Reduce the max number of QPs to 128, which supports standard descriptors
numbers, including the 4K number provided in the test.

Fixes: 80eded877d7c ("crypto/mlx5: support queue pairs operations")

Signed-off-by: Raja Zidane <rzidane@nvidia.com>