app/testpmd: cleanup shared flow resources after port close or detach
When a port was closed or detached (and implicitly closed by the PMD),
flow indirect actions could remain with their handles no longer valid.
If a newly attached device was assigned the same ID as the closed port,
those indirect actions became accessible again.
Any attempt to use them resulted in an undefined behavior.

Flex items had no such issue on close, but had it on detach.

Introduce flush_port_owned_resources() function
for consistent cleanup and call it when a port is closed or detached.

Fixes: 55509e3a49fb ("app/testpmd: support shared flow action")
Fixes: 783a6ea01484 ("app/testpmd: add flex item commands")

Signed-off-by: Dmitry Kozlyuk <dkozlyuk@nvidia.com>
Acked-by: Matan Azrad <matan@nvidia.com>