net/mlx5: fix shared Rx queue segment configuration match
[ upstream commit 572c9d4bda08555f21ad1f4964a60810c1a23d2d ]

While joining the shared Rx queue to the existing queue group,
the queue configurations is checked to be the same as it was
specified in the first group queue creation - all shared
queues should be created with identical configurations.

During the Rx queue creation the buffer split segment
configuration can be altered - the zero segment sizes are
substituted with the actual ones, inherited from the pools,
number of segments can be extended to cover the maximal
packet length, etc. It means the actual queue segment
configuration can not be used directly to match the
configuration provided in the queue setup call.

To resolve an issue we should store original parameters
in the shared queue structure and perform the check against
one of these stored ones.

Fixes: 5a1d58468a59 ("net/mlx5: fix multiple Rx segment check for shared Rx queue")

Signed-off-by: Viacheslav Ovsiienko <viacheslavo@nvidia.com>

Signed-off-by: Xiaoyu Min <jackmin@nvidia.com>