net/mlx5: fix matching with VLAN
In commit [1] all the match items are split to mask and key separately.
The change introduced a bug which sets the VLAN tag incorrectly in the
mask. The bug caused the VLAN match pattern to be invalid.

This commit adds the protection to set the VLAN tag only in key mode
as before.

Fixes: c64b5a27b3f2 ("net/mlx5: split Direct Verbs flow item translate")

Signed-off-by: Suanming Mou <suanmingm@nvidia.com>