net/mlx5: fix matching with VXLAN
For the NIC which doesn't support matching on MISC5 with non-default
UDP port 4789, matching on MISC is preferred.

In the previous HWS change [1] all match items are split to mask and
key separately. The change introduced a bug which caused the UDP port
in mask case can't be get properly.

This commit adds the UDP port to the matcher workspace to avoid the
incorrect UDP port read in VXLAN mask configure routine.

Fixes: c64b5a27b3f2 ("net/mlx5: split Direct Verbs flow item translate")

Signed-off-by: Suanming Mou <suanmingm@nvidia.com>