net/mlx5: fix matcher mask zero of metadata matching
The flow_dv_match_meta_reg() is used for converting both the value
and mask in flow_dv_match_meta_reg_all(). To remove the irrelevant
bits in the value, an bitwise AND operation is used inside.

When converting the mask, another mask must be used as the input
parameter in case that the 0 mask is used after the AND.

Fixes: c64b5a27b3f2 ("net/mlx5: split Direct Verbs flow item translate")

Signed-off-by: Bing Zhao <bingz@nvidia.com>