net/mlx5: fix unwanted sft stop
The "sft_en" in the structure of "mlx5_priv" is used to indicate that
sft function driver is started. When stopping a device, this flag
should be used to decide if sft driver needs to be stopped.

Another flag in the "mlx5_dev_config" is used to indicate if this
device support sft function.

The misuse of this flag in "mlx5_dev_config" caused an unwanted sft
stop during device stop. It caused an overflow of the mark flag in a
queue and an incoming packet's mark could not be retrieved in the
datapath.

In this commit, the correct flag is used.

Fixes: 35cb11cb6977 ("net/mlx5: fix SFT rules list assignments")

Signed-off-by: Bing Zhao <bingz@nvidia.com>