net/mlx5: fix the missing set of ethertype in mask
The function flow_dv_translate_item_eth() had a refactor. The mask
and value's converting calls this twice and they are handled
independently.

When converting the mask, if both the mask of value has_vlan is true,
then there is no need to match the ethertype field. The mask of this
field will be skipped and handled in the VLAN converting.

In other cases:
  1. has_vlan in mask is 0: the ethertype should be matched and the
     field should be filled.
  2. has_vlan in mask is 1 but in value is 0: there should not be any
     VLAN of the incoming packets, the ethertype should be filled.

Fixes: c64b5a27b3f2 ("net/mlx5: split Direct Verbs flow item translate")

Signed-off-by: Bing Zhao <bingz@nvidia.com>