net/mlx5: fix indexed pool fetch overlap issue
For indexed pool with local cache, when a new trunk is allocated,
currently half of the trunk's index will be fetched to the local
cache. In case the local cache size is less then half of the trunk
size, memory overlap happened.

This commit adds the check of the fetch size, if local cache size
is less than fetch size, adjust the fetch size to be local cache
size.

Fixes: ecadd7ce40c4 ("net/mlx5: add indexed pool local cache")

Signed-off-by: Suanming Mou <suanmingm@nvidia.com>