common/mlx5: add list registration lookup retry
During core shared list registration, if entry is not exist in
local list, multi-thread may try to create the same entry at the
same time. Some of the lists does not support create the same
duplicate entry. In that case the registration will be failed
at once.

This commit adds a lookup retry from global list to avoid the
failure.

Signed-off-by: Suanming Mou <suanmingm@nvidia.com>