app/testpmd: fix incorrect action end
The HW steering action template added the template mask at the end of
common action. For the commands such as meter policy requires multiple
next_action end with the first policy action.

This commit fixes the bug with changing the action template mask as the
sub-command of template ID instead of starting from common action end.

Fixes: 32aa1edba851 ("app/testpmd: implement rte flow item/action template")

Signed-off-by: Suanming Mou <suanmingm@nvidia.com>