net/mlx5: fix mark action mask
Currently, for actions in the template with mask conf not NULL, means
the action will be a constant action. The mark action hadn't checked
the mask conf value in translate which caused the tag value will be
updated always with the flow tag value.

This commit adds the check of mark action mask to configure the tag
value in the action accordingly.

Fixes: 80be4cfe2630 ("net/mlx5: add mark action")

Signed-off-by: Suanming Mou <suanmingm@nvidia.com>