mlx5dr: Send, Provide queue action support
This allows to trigger a start drain action on the queue.
When queuing a create or destroy rule operation the execution
is not done immediately since we allow accumulating a burst of
operations. The new DRAIN action allows to signal the queue to
start executing all pending queued rules and write to the HW.
The new API may allow a combination of actions on the queue.

Signed-off-by: Alex Vesker <valex@nvidia.com>