mlx5dr: FDB support
FDB is implemented using FDB RX and FDB TX both
managed by the SW. The insertion logic to RX or TX
is managed by the mlx5dr API. To reduce the complexity
oh splitting the code to RX and TX the FDB abstraction
is done only in the resource allocation layer.
-Once we allocate an FDB resource such as STC/STE
 we allocate two resources FDB_RX and FDB_TX which
 are managed and abstracted in pool layer
-From mlx5dr point of view we use a single resource
 for performance the same index is used in both RX and TX

This design allows for clear code with high performance
since only a single FDB resource is managed.

Signed-off-by: Alex Vesker <valex@nvidia.com>