mlx5dr: Action, fix bug in one-action modify-header
The data that belong to the one-action should come from the bulk or from
the action itself if it was in SHARED mode.
now, when the data is in the bulk it contains 8B, 4 for the pattern and
only the last for are for the data.
When the data from the action we should move it to big endian, because
it was kept as little endian.

Signed-off-by: Erez Shitrit <erezsh@nvidia.com>