mlx5dr: Definer, Avoid accessing invalid selectors
There can be an issue with the selector value 0, since we
don't have a reserved value for selectors, on definer binding
we can take an incorrect selector. This is not critical for
DWs but it can lead to an issue with BYTEs.

For example when searching for an SMAC (offset 0), we would
try to find a byte selector with value offset 0, the code
assumes that fields are continues across bytes.

To resolve this we will:
-Search in DW first
-If not found in DW will reverse search in bytes,
 Search is in reverse since we fill the bytes in reverse
-The search byte offset MUST be present

Signed-off-by: Alex Vesker <valex@nvidia.com>