app/compress-perf: fix cyclecount enqueue ops allocation
In cyclecount main_loop function, each iteration it tries to enqueue X ops,
in case Y<X ops were enqueued, the rest of the X-Y ops are moved to the
beginning of the ops array, to preserve ops order, and next Y ops are
allocated for the next enqueue action, the allocation of the ops occurs on
the first Y entries in the array, when it should have skipped the first X-Y
array entries and allocate the following Y entries.

Fix the allocation by adding the correct offset.

Fixes: 2695db95a147 ("test/compress: add cycle-count mode to perf tool")
Cc: stable@dpdk.org

Signed-off-by: Raja Zidane <rzidane@nvidia.com>