net/mlx5: disable limited flow indexed pool local cache
HW steering achieves the high insertion rate with indexed pool
local cache and resource bulk allocation.

But that only makes sense when inserting huge number of flows
such as millions of flows to the table. Since bulk resources
management gets rid of slow allocation for every flow insertion.
For the tables only create limited number of flows, insertion
rate will still be very limited as the resource will be used out
shortly and insertion rate will be lag. And in that limited flow
case, the index be cached and separated in multi-core local cache
become big weight to the total flow number. That causes the flow
allocation failed with not enough memory as the other cores cache
the index to their own local cache.

This commit disables the limited flow index pool local cache to
avoid flow allocation fail with not enough memory.

Signed-off-by: Suanming Mou <suanmingm@nvidia.com>