mlx5dr: Action, Use the same action type for single modify action
This is done since we want to expose action template which requires
the upper layer to pass the action type. The SET/COPY/ADD is an
internal optimization, we cannot expect upper layer to be aware
of it. The other option is to have two mapping internal and external
enum, I preferred to avoid it.

I am hoping that the final template action will align all of the ifs.

Signed-off-by: Alex Vesker <valex@nvidia.com>