net/mlx5: add default Tx metadata copy rule
When opening the context and creating the queues, the default table
and rule will be inserted for HWS if the copy of Tx metadata is
specified from the devarg.

The default rule is to copy Tx metadata to REG_C_1 for all packets.
It is with a low priority. If there is other Tx rule needed and a
jump from root to the next tables is executed, this rule will be
bypassed.

Signed-off-by: Bing Zhao <bingz@nvidia.com>