net/mlx5: create port actions
This patch implements creating and caching of port actions for use with
HW Steering FDB flows.

Actions are created on flow template API configuration and created
only on the port designated as master. Attaching and detaching of ports
in the same switching domain causes an update to the port actions cache
by, respectively, creating and destroying actions.

Signed-off-by: Dariusz Sosnowski <dsosnowski@nvidia.com>