net/mlx5: use fixed shared action of header modify
The header modify action shouldn't be added into the actions list
when compiling. The fixed action will be used and there is no need
to construct another rte_flow_action* array with the implicit fixed
copy action for Rx metadata.

The source action position will be adjusted(-1) if there is an
implicit modify header inserted before the Q/RSS action.

Signed-off-by: Bing Zhao <bingz@nvidia.com>