net/mlx5: enable mark flag for all ports in the same domain
In the switchdev mode, there is a unique FDB domain for all the
representors and only the eswitch manager can insert the rule into
this domain.

If a flow rule is like below:
flow create 0 ingress transfer pattern port_id id is X / eth / end
actions mark id 25 ...
It is used for representor X and the mark flag was not enabled for
the queues of this port.

To fix this, once the mark flag needs to be enabled, in a FDB case,
all the queues' mark flag belonging to the same domain will be
engaged for only once.

Fixes: e211aca851a7 ("net/mlx5: fix mark enabling for Rx")

Signed-off-by: Bing Zhao <bingz@nvidia.com>