net/mlx5: fix the unneeded source port match
The NIC Rx domain is also shared among all the representors inside
the same eswitch domain. The internal rule in the MREG copy table is
used to match on the MARK in REG_C_x and set the MARK to flow_tag
field again without changing other fields.

It should be global and there is no need to match the source port.

Fixes: dd3c774f6ffb ("net/mlx5: add metadata register copy table")

Signed-off-by: Bing Zhao <bingz@nvidia.com>