net/mlx5: fix the shared modify header content
Once a shared modify header is created per table, it will be applied
to all the flow rules created on this table. The steering action
object and its content is filled when creating the action.

The content part is a union and resetting it to empty would cause the
action cannot be used in the steering and there is no modify header
taking effect.

Fixes: 234752a5249c ("net/mlx5: add modify field hws support")

Signed-off-by: Bing Zhao <bingz@nvidia.com>