net/mlx5: fix represented port matching on HWS
On SW Steering a hint was provided to the matcher when flow was matching
on the uplink. For uplink, if vport's metadata tag was available, flow
was matched on both metadata tag and source vport.

In case of HW Steering matching for uplink and VF/SF vports is realized
using only metadata tag. This patch disables the aforementioned hint
when HW Steering is engaged, so that matcher is configured to match only
on metadata tags.

Fixes: bd5c8216c2a7 ("net/mlx5: support represented port item in HWS")

Signed-off-by: Dariusz Sosnowski <dsosnowski@nvidia.com>