net/mlx5: add action mask value checking
As the rte_flow_async API document describes, for the actions with
conf field as value, value zero means the action in action template
should refer the conf detail and be created individually as the flow
indicates. Non zero means the action will be created as the template
indicates.

This commit adds the action mask value checking for such cases.

Signed-off-by: Suanming Mou <suanmingm@nvidia.com>