net/mlx5: fix the flow source vport checking in validation
The flow source vport was checked according to the port id in
the flow creation. While the port id match was added in the flow
match, the actual source vport was changed.

This patch adds the new function to check the actual flow source
port is representor port or not. And updates the PUSH VLAN, POP VLAN,
and flow sampler action validation.

Fixes: e10d04e5b53b ("net/mlx5: fix VLAN push action validation")
Fixes: 27d2b064cf19 ("net/mlx5: add check for pop and push VLAN actions")
Fixes: f2e7bc5cd842 ("net/mlx5: fix E-Switch egress mirror flow validation")
Cc: stable@dpdk.org

Signed-off-by: Jiawei Wang <jiaweiw@nvidia.com>