net/mlx5: fix tag register retrieve in the steering
The function mlx5_flow_field_id_to_modify_info() is shared between
the software steering(SWS) and hardware steering(HWS) and it is used
to construct the information of a header modify action. Regarding
the field of "RTE_FLOW_FIELD_TAG", the available tag registers are
different in HWS from SWS.

The devarg of "dv_flow_en" should be checked in order to provide the
correct information of a REG_C_x for tag.

Fixes: 43df7be5f2f7 ("net/mlx5: provide the available tag registers")

Signed-off-by: Bing Zhao <bingz@nvidia.com>