net/mlx5: fix group translation for jump actions
Commit 3d305bd2b019 ("net/mlx5: translate group for transfer flows in HWS")
introduced translation of flow groups for transfer flows. With that
commit group passed in flow table attributes was translated to higher
group. However the destination group of jump actions was not adequately
translated.

This patch fixes this behavior. It adds translation of destination
groups in jump actions. Groups are translated if and only if:

- E-Switch is enabled,
- flow table is created through flow API,
- flow tables contains transfer flows.

Information required to check if translation is needed is stored in
rte_flow_template_table struct on table creation.

Fixes: 3d305bd2b019 ("net/mlx5: translate group for transfer flows in HWS")

Signed-off-by: Dariusz Sosnowski <dsosnowski@nvidia.com>