net/mlx5: query hws counter on DCS basis
Now the counter id is optimized by using DCS idx as part of it, which
leads to fetch dr action more optimized.

For the ASO query, we also can use this benefit to compose WQE in more
efficient way, which is to query counter on DCS basis.

In this way, we know the start DCS id value of this bulk and id should
be continusely growth. During WQE composing, we just simply needs to
increase DCS id, needn't to go through all DCSs in order to find the id.

Signed-off-by: Xiaoyu Min <jackmin@nvidia.com>