net/mlx5: remove redundant proxy port log
mlx5 PMD reported that it was "Unable to find proxy port" while
configuring a port. This was reported during flushing of HW Steering
control flows. HW Steering control flows are stored on the flow list
associated with transfer proxy port. Without E-Switch enabled PMD is
unable to find transfer proxy port and it cannot access the relevant
flow list.

This patch fixes this behavior by looking for transfer proxy port if and
only if E-Switch is enabled. If E-Switch is disabled, then the flow list
on the given port is used.

Signed-off-by: Dariusz Sosnowski <dsosnowski@nvidia.com>