pydiru: adjust Cython to support static global C vars
Since pydiru uses internal (mlx5 HWS) functions it must be compiled
against the relevant static libraries.
But pydiru tests are in python, hence pydiru is built as shared
libraries, which creates issues with the statically linked libraries in
some cases, for example when using global (extern) variables in C that
are shared among different modules or C files.
To solve that issue, a new main mlx5dr Cython module is introduced in
pydiru that wraps all the mlx5dr library functions in order to create
one shared library, and only this library uses the internal functions
directly.

Signed-off-by: Daria Velikovsky <daria@nvidia.com>
Signed-off-by: Edward Srouji <edwards@nvidia.com>