net/mlx5: fix wrong group value for meter suffix flow
While the fdb_def_rule_en was set to 0, the dedicated E-Switch rule
was not added to the root table, and the flow group value of external
transfer flows were kept at the original value.

This patch adds the checking for meter suffix flow to the user the correct
group value if fdb_def_rule_en is 0.

Fixes: 9736da4d0e25 ("net/mlx5: add the device control for E-Switch default rule")

Signed-off-by: Jiawei Wang <jiaweiw@nvidia.com>