net/mlx5: fix query invalid hws counter
When query HW steering counter, the counter id in rte_flow struct is
used.

However, the check of if this counter id is valid is missing which
causes the flow with an invalid counter id (i.e, flow rule without
counter action) will be always taken as counter id is zero, so the first
counter is returned.

Fix the above by add checking if the counter id is valid.

Fixes: 239d0c1e5762 ("net/mlx5: support HW steering count action")

Signed-off-by: Xiaoyu Min <jackmin@nvidia.com>