vdpa/mlx5: fix device cache clean
The close device operation is done in the background thread
from the MT mechanism in the driver.
The device clean operation can be triggered after closing
when the background thread is in progress.
Wrongly, in this case, the clean handler in the driver doesn't wait for
the close to the finish, which may cause synchronization issues.

Fixes: 3c936ad8ff9e ("vdpa/mlx5: add device close task")
Cc: stable@dpdk.org

Signed-off-by: Li Zhang <lizh@nvidia.com>