net/mlx5: query max used hws counter
Current PMD queries all HW steering counters configured by the user.
However, users may use some of them increasingly. In this case, querying
all the counters wastes CPU cycles, which is unnecessary.

This patch will query counter up to max used counter till now, which
means even users release some counters after max counters are used, still
the max counter will be queried. In other words, only track increased
the number of counters in use, not the number of decreased counter
in use.

In order to track the max number of counters in use, an additional
reuse_list is created, which is also a rte_ring with the same size of
the global free list.

When the user creates a counter, bring it from the global reuse list if
the cache is empty instead of fetching it from the global free list.
If the reuse list is empty, fetch from the global free list.

For the background service thread, which will move the counters in the
global reset list into the global reuse list instead of the international
free list.

Before query counter, check how many counters are still in the free list,
means they are not used and only query the number of counters which is
used.

Signed-off-by: Xiaoyu Min <jackmin@nvidia.com>