net/mlx5: fix retrieving tag item in root table
In the hardware steering(HWS), the matcher creations from items for
root and non-root tables are different. In the root table, it will
reuse the shared code from software steering.

Since the available tag registers are different in HWS from SWS, a
check is needed to use the correct function to retrieve the correct
REG_C_x from the index.

Fixes: 43df7be5f2f7 ("net/mlx5: provide the available tag registers")

Signed-off-by: Bing Zhao <bingz@nvidia.com>