net/mlx5: fix headroom corruption in case of small stride
There are two ways to handle packet data in case of MPRQ.
1. Copy packet data to user mbuf if the packet is small.
2. Attach MPRQ buffer with packet data to user mbuf otherwise.
MPRQ reserves 128 bytes from the tail of a preceding stride
to provide headroom for user mbufs in the second case.
External mbuf is pointed to this address to keep the headroom.
And this chunk of packet data is copied to user mbuf.

But MPRQ stride size can be configured as small as 64 bytes.
That means we have to clear the previous packet as well in
order to preserve headroom data, which belongs to a user already.
Or users may request packet data copy size to not exceed a certain
size preventing PMD from properly securing the headroom.

Check these conditions and set the stride size to the headroom size
in case there is a configuration that leads to headroom corruption.

Fixes: ecb160456a ("net/mlx5: add device parameter for MPRQ stride size")
Cc: stable@dpdk.org

Signed-off-by: Alexander Kozyrev <akozyrev@nvidia.com>