net/mlx5: fix Tx metadata copy flow rule creation
Before a port is started, no rule is allowed to be inserted into the
device, even for the default or PMD internal rules. The pattern and
action templates could be created before the starting a port.

With this commit, the default Tx metadata copy rule is postponed until
enabling the traffic. The default table will remain in the system
unless there is a device close action.

Fixes: ab964209126e ("net/mlx5: fix create table with RSS action crash")
Fixes: d1d90f363d3e ("net/mlx5: add default Tx metadata copy rule")

Signed-off-by: Bing Zhao <bingz@nvidia.com>