net/mlx5: fix the metadata register usage when matching
The metadata REG_B couldn't be used for matching right now. In the
extended metadata mode, the REG_C_1 is used to set metadata in the
FDB domain and copied to the NIC Rx domain.

In the NIC Rx domain, a copy action was done to copy the content
from REG_C_1 to REG_B. The value in REG_B will be observed in the
CQE eventually. If a matching is also needed, the REG_C_1 can be
used instead of REG_B.

With this commit, the correct register is used for the root table in
the hardware steering.

Fixes: 43df7be5f2f7 ("net/mlx5: provide the available tag registers")

Signed-off-by: Bing Zhao <bingz@nvidia.com>