mlx5dr: Definer, Fix BE setters
In cases in which we write to unaligned value, for example BE16:
The following DW [0-7, 8-15, 16-23, 24-31] and we want to write
to 8-23 will fail since the pointer only does 16 bits jumps.
Also added a comment about BE32, BE16 functions which require
setting the full 32/16 bit value, the bit_off and mask are
ignored in such case.
BE32 was also fixed although there is no use case in which we
write to an unaligned pointer due to definer and header layout.

Signed-off-by: Alex Vesker <valex@nvidia.com>
Signed-off-by: Muhammad Sammar <muhammads@nvidia.com>