mlx5dr: Definer, Fix binded bit offset calculation
When binding to definer we set the new bit offset based
on the location in the definer. The calculation was incorrect
and lead to incorrect offset.

For example:
Address   0     1     2     3
Memory: ByteA ByteB ByteC ByteD

If ByteD was binded moved to ByteC we would expect the bit offset
to change from 0 to 8, but the calculation was negative.

Signed-off-by: Alex Vesker <valex@nvidia.com>
Signed-off-by: Muhammad Sammar <muhammads@nvidia.com>