net/mlx5: fix IPv6 and TCP RSS hash fields
In the flow_dv_hashfields_set() function, while item_flags was 0,
the code went directly to the first if and the else case would
never have chance be checked. This caused the IPv6 and TCP hash
fileds in the else case would never be set.

This commit adds the dedicate HW steering hash field set function
to generate the RSS hash fields.

Fixes: 6540da0b93b5 ("net/mlx5: fix RSS scaling issue")
Signed-off-by: Suanming Mou <suanmingm@nvidia.com>