app: add option to compile disabled compilation apps
This is useful in case we have app that is disabled by default, and
we want to allow the user to enable this app compilation at the run
time.

Usage example:
APP_X is application in /apps that the compile flag for it is False.
To compile DPDK with this APP_X use:

meson -Dapps=<APP_X> build; ninja -C build

Signed-off-by: Wisam Jaddo <wisamm@nvidia.com>