net/mlx5: fix the available tags for multiple devices
When initializing another device in the same process, the previous
initialized available tags array needs to be checked. A 8-bits mask
is constructed to check the available tags from the previous
initialized devices. The bit should be set to indicate the original
position of the tag in the array.

The previous implemetation wrongly set the mask and would cause the
incorrect information of the available tags. With this commit, the
mask set action is fixed.

Fixes: 43df7be5f2f7 ("net/mlx5: provide the available tag registers")

Signed-off-by: Bing Zhao <bingz@nvidia.com>