net/mlx5: add support for ASO return register
A REG_C_x metadata register is needed to store the result after an
ASO action. Like in the SWS, the meter color register is used for
all the ASO actions right now and this register was already filtered
out from the available tags.

It is assumed that all the devices are using the same meter color
register inside one application now.

In the next stage, the available tags and other metadata registers
allocation will be stored per device.

Signed-off-by: Bing Zhao <bingz@nvidia.com>