net/mlx5: fix wrongly take hws counter as shared
PMD will put HW steering counter back into pool only if the counter
is not shared when user destroy flow rule.
The _share_ bit in counter struct is for this purpose.

However, this _share_ bit is a union with the _query_gen_when_free_
which record counter's query gen when it is released.
So if the _query_gen_when_free_'s lowest bit is 1, for example: 0x3,
It will be wrongly taken as *shared* counter because the _share_ bit is
set and it will not be put back into pool when user destroy the flow
rule.

Fix above by explicitly clear _share_ bit when get counter.

Fixes: 469a8a32d318 ("net/mlx5: support hws counter action with mask")

Signed-off-by: Xiaoyu Min <jackmin@nvidia.com>