net/mlx5: fix double free risk in port configure
The flow_hw_configure() function allocates a temporary variable mamed
"_queue_attr".
This temporary allocation is freed either at the end of the function or
in the error label.

When the counter support was added, its configuration handling was
accidentally added after the "_queue_attr" was freed.
When the configuration fails, it jumps to the error label and releases
it again.

This patch moves the "_queue_attr" to the end of function.

Fixes: 239d0c1e5762 ("net/mlx5: support HW steering count action")
Cc: jackmin@nvidia.com

Signed-off-by: Michael Baum <michaelba@nvidia.com>